build: move the homepage to root
