+++ frontend/src/app/layout.tsx
@@ -1,16 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
+import Header from "./component/header";
+import Sider from "./component/sider";
 
 export const metadata: Metadata = {
   title: "Create Next App",
@@ -23,12 +14,13 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
+    <div className="flex flex-row justify-start w-full gap-4 p-4 ">
+      <Sider />
+      <div lang="en" className="w-full">
+        <Header />
+
         {children}
-      </body>
-    </html>
+      </div>
+    </div>
   );
 }
